LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-volume-container-mute-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing <code>MacOSInlineMediaControls</code> mute button visibility behavior with volume slider container.");

window.jsTestIsAsync = true;

const mediaControls = new MacOSInlineMediaControls({ width: 500, height: 200 });

mediaControls._host = {
    isMediaControlsMacInlineSizeSpecsEnabled: true
};

if (mediaControls.playPauseButton) {
    mediaControls.playPauseButton.enabled = true;
}

mediaControls.showsStartButton = false;
mediaControls.visible = true;

function isVolumeSliderContainerShowing() {
    const children = mediaControls.children || [];
    return children.includes(mediaControls._volumeSliderContainer);
}

function flushAllLayouts() {
    for (let i = 0; i < 3; i++) {
        scheduler.flushScheduledLayoutCallbacks();
    }
}

function testMuteButtonVisibilityBehavior() {
    debug("Testing mute button visibility behavior:");
    
    // Test 1: Volume container shows - mute button should be visible
    mediaControls.needsLayout = true;
    flushAllLayouts();
    
    shouldBeTrue("isVolumeSliderContainerShowing()", "Volume slider container should be showing");
    shouldBeTrue("mediaControls.muteButton.visible", "Mute button should be visible when container shows");
    
    // Test 2: Width < 60 - special case forces mute button hidden
    mediaControls.width = 10;
    mediaControls.needsLayout = true;
    flushAllLayouts();
    
    shouldBeFalse("isVolumeSliderContainerShowing()", "Volume slider container should be hidden");
    shouldBeFalse("mediaControls.muteButton.visible", "Mute button should be hidden when width < 60 (forced)");
    
    // Test 3: Restore width - mute button becomes visible again
    mediaControls.width = 500;
    mediaControls.needsLayout = true;
    flushAllLayouts();
    
    shouldBeTrue("isVolumeSliderContainerShowing()", "Volume slider container should show again");
    shouldBeTrue("mediaControls.muteButton.visible", "Mute button should be visible when container is restored");
    
    // Test 4: showsStartButton hides container but doesn't force mute button hidden
    mediaControls.showsStartButton = true;
    mediaControls.needsLayout = true;
    flushAllLayouts();
    
    shouldBeFalse("isVolumeSliderContainerShowing()", "Volume slider container should be hidden");
    shouldBeTrue("mediaControls.muteButton.visible", "Mute button visibility is preserved when container is hidden by showsStartButton");
}

function ready() {
    return mediaControls.muteButton && mediaControls.playPauseButton &&
           mediaControls._volumeSliderContainer &&
           mediaControls.muteButton.width > 0 && mediaControls.playPauseButton.width > 0;
}

shouldBecomeEqual("ready()", "true", () => {
    debug("");
    testMuteButtonVisibilityBehavior();
    debug("");
    finishJSTest();
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
