LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-volume-container-mute-button-expected.txt

Testing MacOSInlineMediaControls mute button visibility behavior with volume slider container.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ready() became true

Testing mute button visibility behavior:
PASS isVolumeSliderContainerShowing() is true
PASS mediaControls.muteButton.visible is true
PASS isVolumeSliderContainerShowing() is false
PASS mediaControls.muteButton.visible is false
PASS isVolumeSliderContainerShowing() is true
PASS mediaControls.muteButton.visible is true
PASS isVolumeSliderContainerShowing() is false
PASS mediaControls.muteButton.visible is true

PASS successfullyParsed is true

TEST COMPLETE
